[Bronze V] Title: 사파리월드, Time: 172 ms, Memory: 17632 KB -BaekjoonHub

+++ 백준/Bronze/2420. 사파리월드/사파리월드.java
@@ -5,9 +5,9 @@ import static java.lang.Math.abs;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long N = sc.nextInt();
-        long M = sc.nextInt();
+        int N = sc.nextInt();
+        int M = sc.nextInt();
 
-        System.out.println(abs(N-M));
+        System.out.println(abs((long)N-M));
     }
 }
